parametrize verbose test

+++ pytest_fixtures/component/templatesync.py
@@ -42,9 +42,9 @@ def git_port(default_sat):
 def git_pub_key(default_sat, git_port):
     """Copy ssh public key to git service"""
     git = settings.git
-    key_path = '/usr/share/foreman/.ssh/'
-    default_sat.execute(f'sudo -u foreman ssh-keygen -q -t rsa -f {key_path}id_rsa -N "" <<<y')
-    key = default_sat.execute(f'cat {key_path}id_rsa.pub').stdout.strip()
+    key_path = '/usr/share/foreman/.ssh'
+    default_sat.execute(f'sudo -u foreman ssh-keygen -q -t rsa -f {key_path}/id_rsa -N "" <<<y')
+    key = default_sat.execute(f'cat {key_path}/id_rsa.pub').stdout.strip()
     title = gen_string('alpha')
     auth = (git.username, git.password)
     url = f'http://{git.hostname}:{git.http_port}'
@@ -67,28 +67,11 @@ def git_pub_key(default_sat, git_port):
     res.raise_for_status()
 
 
-@pytest.fixture(scope='session')
-def git_repository(git_port, git_pub_key):
-    """Encapsulate creating initialized repository"""
-    repo_gen = create_git_repository(True)
-    yield next(repo_gen)
-    next(repo_gen)
-
-
 @pytest.fixture(scope='function')
-def git_empty_repository(git_port, git_pub_key):
-    """Encapsulate creating empty repository"""
-    repo_gen = create_git_repository(False)
-    yield next(repo_gen)
-    next(repo_gen)
-
-
-def create_git_repository(init):
+def git_repository(git_port, git_pub_key, request):
     """Creates a new repository on git provider for exporting templates.
 
     Finally, deletes repository from git provider after tests are completed as a teardown part.
-
-    :returns: dict: repository name, flag if repository is empty
     """
     auth = (settings.git.username, settings.git.password)
     url = f'http://{settings.git.hostname}:{settings.git.http_port}'
@@ -96,13 +79,12 @@ def create_git_repository(init):
     res = requests.post(
         f'{url}/api/v1/user/repos',
         auth=auth,
-        json={'name': name, 'auto_init': init, 'default_branch': 'master'},
+        json={'name': name, 'auto_init': request.param, 'default_branch': 'master'},
     )
     res.raise_for_status()
-    yield name
+    yield {'name': name, 'init': request.param}
     res = requests.delete(f'{url}/api/v1/repos/{settings.git.username}/{name}', auth=auth)
     res.raise_for_status()
-    yield
 
 
 @pytest.fixture()
@@ -111,14 +93,17 @@ def git_branch(git_repository):
 
     Finally, removes branch from the git repository after test is completed as a teardown part.
     """
-    auth = (settings.git.username, settings.git.password)
-    path = f'/api/v1/repos/{settings.git.username}/{git_repository}/branches'
-    url = f'http://{settings.git.hostname}:{settings.git.http_port}{path}'
-    new_branch = gen_string('alpha')
-    res = requests.post(
-        url, auth=auth, json={'new_branch_name': new_branch, 'old_branch_name': 'master'}
-    )
-    res.raise_for_status()
-    yield new_branch
-    res = requests.delete(f'{url}/{new_branch}', auth=auth)
-    res.raise_for_status()
+    if git_repository['init']:
+        auth = (settings.git.username, settings.git.password)
+        path = f'/api/v1/repos/{settings.git.username}/{git_repository["name"]}/branches'
+        url = f'http://{settings.git.hostname}:{settings.git.http_port}{path}'
+        new_branch = gen_string('alpha')
+        res = requests.post(
+            url, auth=auth, json={'new_branch_name': new_branch, 'old_branch_name': 'master'}
+        )
+        res.raise_for_status()
+        yield new_branch
+        res = requests.delete(f'{url}/{new_branch}', auth=auth)
+        res.raise_for_status()
+    else:
+        yield 'master'
